+++ allure_results/0d2d62e3-0fd5-42d2-93eb-1de860400d99-result.json
@@ -1,0 +1,1 @@
+{"name": "Проверка заказа с помощью нижней кнопки Заказать", "status": "passed", "steps": [{"name": "Нажать на нижнюю кнопку Заказать", "status": "passed", "start": 1686750310277, "stop": 1686750310579}, {"name": "Ввод имени и фамилии", "status": "passed", "start": 1686750310579, "stop": 1686750310836}, {"name": "Ввод улицы + выбор метро", "status": "passed", "start": 1686750310836, "stop": 1686750312001}, {"name": "Ввод телефона", "status": "passed", "start": 1686750312001, "stop": 1686750312087}, {"name": "Нажать на кнопку Далее", "status": "passed", "start": 1686750312087, "stop": 1686750312344}, {"name": "Заполняем дату и срок аренды", "status": "passed", "start": 1686750312344, "stop": 1686750313882}, {"name": "Выбираем цвет и оставляем комментарий", "status": "passed", "start": 1686750313882, "stop": 1686750314235}, {"name": "Нажимаем кнопку Заказать + кнопку Да", "status": "passed", "start": 1686750314235, "stop": 1686750314746}], "start": 1686750308358, "stop": 1686750314788, "uuid": "840814fd-05cd-42eb-8718-69d853d3e98d", "historyId": "f442946af2a7cc32cdcd270443872784", "testCaseId": "f442946af2a7cc32cdcd270443872784", "fullName": "test.test_order.TestOrders#test_checkout_low_button", "labels": [{"name": "parentSuite", "value": "test"}, {"name": "suite", "value": "test_order"}, {"name": "subSuite", "value": "TestOrders"}, {"name": "host", "value": "LAPTOP-E40EENQO"}, {"name": "thread", "value": "15948-MainThread"}, {"name": "framework", "value": "pytest"}, {"name": "language", "value": "cpython3"}, {"name": "package", "value": "test.test_order"}]}
