+++ allure_results/20c8bfb5-e0ec-46c3-82dc-0dfe6859d333-container.json
@@ -1,0 +1,1 @@
+{"uuid": "bc8b7452-4b31-4430-a9f1-ed45f506b402", "children": ["2021e47c-0b94-43f3-a1a7-2ba1bc18e727"], "befores": [{"name": "driver", "status": "passed", "start": 1686750280259, "stop": 1686750287652}], "afters": [{"name": "driver::0", "status": "passed", "start": 1686750294373, "stop": 1686750297574}], "start": 1686750280259, "stop": 1686750297574}
