+++ allure_results/21d9173b-0560-4d3e-8c01-9399e14601f7-container.json
@@ -1,0 +1,1 @@
+{"uuid": "1cc8fb6a-ac6d-4988-adcd-fb8bc8921435", "children": ["b3887e93-8a7d-44aa-b567-5e4fa8a8cae5"], "befores": [{"name": "driver", "status": "passed", "start": 1686749789858, "stop": 1686749796683}], "afters": [{"name": "driver::0", "status": "passed", "start": 1686749799365, "stop": 1686749802462}], "start": 1686749789858, "stop": 1686749802466}
