+++ allure_results/62286d82-b2f3-4105-8d91-483f07b1ba00-container.json
@@ -1,0 +1,1 @@
+{"uuid": "70d07b66-61bd-44d7-bb2b-89b1a70cc161", "children": ["80851771-6530-4de2-b741-e59ef312e866"], "befores": [{"name": "driver", "status": "passed", "start": 1686749806514, "stop": 1686749813241}], "afters": [{"name": "driver::0", "status": "passed", "start": 1686749815748, "stop": 1686749818732}], "start": 1686749806514, "stop": 1686749818732}
