+++ allure_results/6a7be02a-cdf8-42d4-9b9c-099cd09d5540-container.json
@@ -1,0 +1,1 @@
+{"uuid": "64756257-7e6f-451c-a9ae-8a91370bdd62", "children": ["840814fd-05cd-42eb-8718-69d853d3e98d"], "befores": [{"name": "driver", "status": "passed", "start": 1686750301632, "stop": 1686750308358}], "afters": [{"name": "driver::0", "status": "passed", "start": 1686750314792, "stop": 1686750318019}], "start": 1686750301632, "stop": 1686750318019}
