+++ allure_results/7d89550a-9060-4e9e-bb9f-d20a277980aa-result.json
@@ -1,0 +1,1 @@
+{"name": "Проверка перехода по ссылке(яндекс) через заказ", "status": "passed", "steps": [{"name": "Нажать на верхнюю кнопку Заказать", "status": "passed", "start": 1686749815293, "stop": 1686749815603}, {"name": "Нажать на кнопку Яндекс", "status": "passed", "start": 1686749815603, "stop": 1686749815735}], "start": 1686749813242, "stop": 1686749815747, "uuid": "80851771-6530-4de2-b741-e59ef312e866", "historyId": "d076df60ebc812f6b36ae86ee97d9199", "testCaseId": "d076df60ebc812f6b36ae86ee97d9199", "fullName": "test.test_link.TestLink#test_link_yandex", "labels": [{"name": "parentSuite", "value": "test"}, {"name": "suite", "value": "test_link"}, {"name": "subSuite", "value": "TestLink"}, {"name": "host", "value": "LAPTOP-E40EENQO"}, {"name": "thread", "value": "12588-MainThread"}, {"name": "framework", "value": "pytest"}, {"name": "language", "value": "cpython3"}, {"name": "package", "value": "test.test_link"}]}
