+++ allure_results/83a7a20a-5627-4687-8e4e-a77e833c4a8a-result.json
@@ -1,0 +1,1 @@
+{"name": "Проверка перехода по ссылке(самокат) через заказ", "status": "passed", "steps": [{"name": "Нажать на верхнюю кнопку Заказать", "status": "passed", "start": 1686749798742, "stop": 1686749799031}, {"name": "Нажать на кнопку Самокат", "status": "passed", "start": 1686749799031, "stop": 1686749799353}], "start": 1686749796690, "stop": 1686749799362, "uuid": "b3887e93-8a7d-44aa-b567-5e4fa8a8cae5", "historyId": "c23b8bf34318d816e41498382dd7bc3f", "testCaseId": "c23b8bf34318d816e41498382dd7bc3f", "fullName": "test.test_link.TestLink#test_link_scooter", "labels": [{"name": "parentSuite", "value": "test"}, {"name": "suite", "value": "test_link"}, {"name": "subSuite", "value": "TestLink"}, {"name": "host", "value": "LAPTOP-E40EENQO"}, {"name": "thread", "value": "12588-MainThread"}, {"name": "framework", "value": "pytest"}, {"name": "language", "value": "cpython3"}, {"name": "package", "value": "test.test_link"}]}
