+++ allure_results/b8a1ceec-2a3d-40e2-ad09-a5b5322ad5e7-container.json
@@ -1,0 +1,1 @@
+{"uuid": "c125806b-4e14-40ab-b9ce-2f83aa6377fd", "children": ["2021e47c-0b94-43f3-a1a7-2ba1bc18e727", "840814fd-05cd-42eb-8718-69d853d3e98d"], "befores": [{"name": "_session_faker", "status": "passed", "start": 1686750280086, "stop": 1686750280259}], "start": 1686750280086, "stop": 1686750318019}
