+++ allure_results/b964c4e5-5e34-407e-9155-6d59ca81cdc6-result.json
@@ -1,0 +1,1 @@
+{"name": "Проверка заказа с помощью верхней кнопки Заказать", "status": "passed", "steps": [{"name": "Нажать на верхнюю кнопку Заказать", "status": "passed", "start": 1686750289802, "stop": 1686750290119}, {"name": "Ввод имени и фамилии", "status": "passed", "start": 1686750290119, "stop": 1686750290379}, {"name": "Ввод улицы + выбор метро", "status": "passed", "start": 1686750290379, "stop": 1686750291654}, {"name": "Ввод телефона", "status": "passed", "start": 1686750291654, "stop": 1686750291745}, {"name": "Нажать на кнопку Далее", "status": "passed", "start": 1686750291745, "stop": 1686750291995}, {"name": "Заполняем дату и срок аренды", "status": "passed", "start": 1686750291995, "stop": 1686750293554}, {"name": "Выбираем цвет и оставляем комментарий", "status": "passed", "start": 1686750293554, "stop": 1686750293847}, {"name": "Нажимаем кнопку Заказать + кнопку Да", "status": "passed", "start": 1686750293848, "stop": 1686750294335}], "start": 1686750287656, "stop": 1686750294373, "uuid": "2021e47c-0b94-43f3-a1a7-2ba1bc18e727", "historyId": "682c49f3e90770c23938f8f60b91edcf", "testCaseId": "682c49f3e90770c23938f8f60b91edcf", "fullName": "test.test_order.TestOrders#test_checkout_top_button", "labels": [{"name": "parentSuite", "value": "test"}, {"name": "suite", "value": "test_order"}, {"name": "subSuite", "value": "TestOrders"}, {"name": "host", "value": "LAPTOP-E40EENQO"}, {"name": "thread", "value": "15948-MainThread"}, {"name": "framework", "value": "pytest"}, {"name": "language", "value": "cpython3"}, {"name": "package", "value": "test.test_order"}]}
