+++ allure_results/c7d4d242-c090-4efc-8d34-ed1da6c1c38d-container.json
@@ -1,0 +1,1 @@
+{"uuid": "39ac4770-8927-4ec7-b730-28a1603d3d50", "children": ["b3887e93-8a7d-44aa-b567-5e4fa8a8cae5", "80851771-6530-4de2-b741-e59ef312e866"], "befores": [{"name": "_session_faker", "status": "passed", "start": 1686749789650, "stop": 1686749789858}], "start": 1686749789650, "stop": 1686749818733}
